Struct pod::endian::EndianPrimitive

pub struct EndianPrimitive<B, T> {
    // some fields omitted
}

A POD container for a primitive that stores a value in the specified endianness in memory, and transforms on get/set

Methods

impl<B: ByteOrder, T: EndianConvert> EndianPrimitive<B, T>

fn new(v: T) -> Self

Creates a new value

fn get(&self) -> T

Transforms to the native value

fn set(&mut self, v: T)

Transforms from a native value