Trait pod::packed::Packed

pub unsafe trait Packed: Unaligned { }

A marker trait indicating that a type is #[repr(packed)].

This means that all its members are packed or have an alignment of 1, and its memory layout is guaranteed to be in member declaration order.