Trait pod::Pod

pub trait Pod: Sized { }

A marker trait indicating that a type is Plain Old Data.

impl this on POD structs to access the slice helper methods of PodExt and I/O helpers Encode and Decode.

#[repr(packed)] and #[repr(C)] may be helpful for consistent representation across platforms.