Trait pod::PodExt

pub trait PodExt: Sized {
    fn as_slice<'a>(&'a self) -> &'a [u8] { ... }
    fn mut_slice<'a>(&'a mut self) -> &'a mut [u8] { ... }
    fn from_slice<'a>(slice: &'a [u8]) -> &'a Self where Self: Unaligned { ... }
    fn from_mut_slice<'a>(slice: &'a mut [u8]) -> &'a mut Self where Self: Unaligned { ... }
    fn from_vec(vec: Vec<u8>) -> Box<Self> where Self: Unaligned { ... }
    fn from_box(slice: Box<[u8]>) -> Box<Self> where Self: Unaligned { ... }
    fn to_vec(self: Box<Self>) -> Vec<u8> { ... }
    fn to_boxed_slice(self: Box<Self>) -> Box<[u8]> { ... }
    fn map<'a, T: Pod + Unaligned>(&'a self) -> &'a T where Self: Unaligned { ... }
    fn map_mut<'a, T: Pod + Unaligned>(&'a mut self) -> &'a mut T where Self: Unaligned { ... }
    unsafe fn uninitialized() -> Self { ... }
    fn zeroed() -> Self { ... }
}

Helper methods for converting POD types to/from byte slices and vectors

Provided Methods

fn as_slice<'a>(&'a self) -> &'a [u8]

Borrows the POD as a byte slice

fn mut_slice<'a>(&'a mut self) -> &'a mut [u8]

Borrows the POD as a mutable byte slice

fn from_slice<'a>(slice: &'a [u8]) -> &'a Self where Self: Unaligned

Borrows a new instance of the POD from a byte slice

Panics

Panics if slice.len() is not the same as the type's size

fn from_mut_slice<'a>(slice: &'a mut [u8]) -> &'a mut Self where Self: Unaligned

Borrows a mutable instance of the POD from a mutable byte slice

Panics

Panics if slice.len() is not the same as the type's size

fn from_vec(vec: Vec<u8>) -> Box<Self> where Self: Unaligned

Converts a byte vector to a boxed instance of the POD type

Panics

Panics if vec.len() is not the same as the type's size

fn from_box(slice: Box<[u8]>) -> Box<Self> where Self: Unaligned

Converts a boxed slice to a boxed instance of the POD type

Panics

Panics if slice.len() is not the same as the type's size

fn to_vec(self: Box<Self>) -> Vec<u8>

Converts a boxed POD to a byte vector

fn to_boxed_slice(self: Box<Self>) -> Box<[u8]>

Converts a boxed POD to a boxed slice

fn map<'a, T: Pod + Unaligned>(&'a self) -> &'a T where Self: Unaligned

Converts a POD type from one to another of the same size.

Panics

Panics if the two types are not the same size

fn map_mut<'a, T: Pod + Unaligned>(&'a mut self) -> &'a mut T where Self: Unaligned

Converts a POD type from one to another of the same size.

Panics

Panics if the two types are not the same size

unsafe fn uninitialized() -> Self

Generates a new uninitialized instance of a POD type.

fn zeroed() -> Self

Creates a new zeroed instance of a POD type.