Trait pod::Pod

pub unsafe trait Pod: Sized {
    fn as_aligned_mut<T: Pod + Aligned<Unaligned=Self>>(&mut self) -> Option<&mut T> where Self: Copy + Unaligned { ... }
    fn from_unaligned_mut<T: Copy + Unaligned>(s: &mut T) -> Option<&mut Self> where Self: Aligned<Unaligned=T> { ... }
    fn from_unaligned<T: Copy + Unaligned>(s: T) -> Self where Self: Aligned<Unaligned=T> { ... }
    fn as_bytes<'a>(&'a self) -> &'a [u8] { ... }
    fn as_mut_bytes<'a>(&'a mut self) -> &'a mut [u8] { ... }
    fn from_bytes<'a>(slice: &'a [u8]) -> Option<Self> { ... }
    fn from_bytes_ref<'a>(slice: &'a [u8]) -> Option<&'a Self> where Self: Unaligned { ... }
    fn from_bytes_mut<'a>(slice: &'a mut [u8]) -> Option<&'a mut Self> where Self: Unaligned { ... }
    fn from_vec(vec: Vec<u8>) -> Result<Box<Self>, Vec<u8>> where Self: Unaligned { ... }
    fn from_box(slice: Box<[u8]>) -> Result<Box<Self>, Box<[u8]>> where Self: Unaligned { ... }
    fn into_vec(self: Box<Self>) -> Vec<u8> { ... }
    fn into_boxed_slice(self: Box<Self>) -> Box<[u8]> { ... }
    fn map<'a, T: Pod + Unaligned>(&'a self) -> Option<&'a T> where Self: Unaligned { ... }
    fn map_mut<'a, T: Pod + Unaligned>(&'a mut self) -> Option<&'a mut T> where Self: Unaligned { ... }
    unsafe fn uninitialized() -> Self { ... }
    fn zeroed() -> Self { ... }
    fn map_slice<'a, T: Pod + Unaligned>(s: &'a [Self]) -> Option<&'a [T]> where Self: Unaligned { ... }
    fn map_slice_mut<'a, T: Pod + Unaligned>(s: &'a mut [Self]) -> Option<&'a mut [T]> where Self: Unaligned { ... }
}

A marker trait indicating that a type is Plain Old Data.

It is unsafe to impl this manually, use #[derive(Pod)] instead.

Provided Methods

fn as_aligned_mut<T: Pod + Aligned<Unaligned=Self>>(&mut self) -> Option<&mut T> where Self: Copy + Unaligned

Safely borrows the aligned value mutably

See also: Aligned::from_unaligned_mut

fn from_unaligned_mut<T: Copy + Unaligned>(s: &mut T) -> Option<&mut Self> where Self: Aligned<Unaligned=T>

Safely borrows the unaligned value mutably

See also: Aligned::from_unaligned_mut

fn from_unaligned<T: Copy + Unaligned>(s: T) -> Self where Self: Aligned<Unaligned=T>

Safely converts an unaligned value to its aligned equivalent

See also: Aligned::from_unaligned

fn as_bytes<'a>(&'a self) -> &'a [u8]

Borrows the POD as a byte slice

fn as_mut_bytes<'a>(&'a mut self) -> &'a mut [u8]

Borrows the POD as a mutable byte slice

fn from_bytes<'a>(slice: &'a [u8]) -> Option<Self>

Safely creates a POD value from a potentially unaligned slice

Returns None if slice.len() is not the same as the type's size

fn from_bytes_ref<'a>(slice: &'a [u8]) -> Option<&'a Self> where Self: Unaligned

Borrows a new instance of the POD from a byte slice

Returns None if slice.len() is not the same as the type's size

fn from_bytes_mut<'a>(slice: &'a mut [u8]) -> Option<&'a mut Self> where Self: Unaligned

Borrows a mutable instance of the POD from a mutable byte slice

Returns None if slice.len() is not the same as the type's size

fn from_vec(vec: Vec<u8>) -> Result<Box<Self>, Vec<u8>> where Self: Unaligned

Converts a byte vector to a boxed instance of the POD type

Fails if vec.len() is not the same as the type's size

fn from_box(slice: Box<[u8]>) -> Result<Box<Self>, Box<[u8]>> where Self: Unaligned

Converts a boxed slice to a boxed instance of the POD type

Fails if slice.len() is not the same as the type's size

fn into_vec(self: Box<Self>) -> Vec<u8>

Converts a boxed POD to a byte vector

fn into_boxed_slice(self: Box<Self>) -> Box<[u8]>

Converts a boxed POD to a boxed slice

fn map<'a, T: Pod + Unaligned>(&'a self) -> Option<&'a T> where Self: Unaligned

Converts a POD type from one to another of the same size.

Returns None if the two types are not the same size

fn map_mut<'a, T: Pod + Unaligned>(&'a mut self) -> Option<&'a mut T> where Self: Unaligned

Converts a POD type from one to another of the same size.

Returns None if the two types are not the same size

unsafe fn uninitialized() -> Self

Generates a new uninitialized instance of a POD type.

fn zeroed() -> Self

Creates a new zeroed instance of a POD type.

fn map_slice<'a, T: Pod + Unaligned>(s: &'a [Self]) -> Option<&'a [T]> where Self: Unaligned

Maps a POD slice from one type to another

Returns None if the output type does not perfectly fit into the slice.

fn map_slice_mut<'a, T: Pod + Unaligned>(s: &'a mut [Self]) -> Option<&'a mut [T]> where Self: Unaligned

Maps a mutable POD slice from one type to another

Returns None if the output type does not perfectly fit into the slice.